Add new simplified C++ compiler
