+++ compiler/x/cpp/compiler.go
@@ -10,10 +10,9 @@ import (
 	"mochi/parser"
 )
 
-// Compiler is a minimal Mochi -> C++ translator used for tests. It only
-// supports simple features like arithmetic, loops and function calls.
-// Unsupported constructs result in an error so the tests can record the
-// failure.
+// Compiler is a very small Mochi to C++ translator used in tests. It
+// intentionally supports only a subset of the language. Unsupported
+// constructs return an error so that tests can record failures.
 type Compiler struct {
 	buf    bytes.Buffer
 	indent int
@@ -36,30 +35,15 @@ func (c *Compiler) writeIndent() {
 	}
 }
 
-// Compile converts a parsed program to C++.
+// Compile translates the given program into C++ source code.
 func (c *Compiler) Compile(p *parser.Program) ([]byte, error) {
 	c.buf.Reset()
-	c.writeln("// Generated by the Mochi experimental C++ compiler")
+	c.writeln("// Generated by Mochi experimental C++ compiler")
 	c.writeln("#include <bits/stdc++.h>")
 	c.writeln("using namespace std;")
 	c.writeln("")
 
-	// small helpers for type traits
-	c.writeln("template<typename T, typename=void> struct has_size : false_type {};")
-	c.writeln("template<typename T> struct has_size<T, void_t<decltype(declval<T>().size()), decltype(declval<T>()[0])>> : true_type {};")
-	c.writeln("template<typename T, typename=void> struct has_key_type : false_type {};")
-	c.writeln("template<typename T> struct has_key_type<T, void_t<typename T::key_type>> : true_type {};")
-	c.writeln("")
-
-	for _, st := range p.Statements {
-		if st.Type != nil {
-			if err := c.compileTypeDecl(st.Type); err != nil {
-				return nil, err
-			}
-			c.writeln("")
-		}
-	}
-
+	// compile function declarations first
 	for _, st := range p.Statements {
 		if st.Fun != nil {
 			if err := c.compileFun(st.Fun); err != nil {
@@ -69,6 +53,7 @@ func (c *Compiler) Compile(p *parser.Program) ([]byte, error) {
 		}
 	}
 
+	// main body
 	c.writeln("int main() {")
 	c.indent++
 	for _, st := range p.Statements {
@@ -110,172 +95,88 @@ func (c *Compiler) compileFun(fn *parser.FunStmt) error {
 	return nil
 }
 
-func (c *Compiler) compileTypeDecl(td *parser.TypeDecl) error {
-	if len(td.Variants) > 0 {
-		return fmt.Errorf("sum types not supported")
-	}
-	c.writeIndent()
-	c.buf.WriteString("struct ")
-	c.buf.WriteString(td.Name)
-	c.buf.WriteString(" {\n")
-	c.indent++
-	for _, m := range td.Members {
-		if m.Field == nil {
-			return fmt.Errorf("methods in type not supported")
-		}
-		typ, err := c.compileType(m.Field.Type)
-		if err != nil && typ == "auto" {
-			return err
-		}
-		c.writeIndent()
-		c.buf.WriteString(fmt.Sprintf("%s %s;\n", typ, m.Field.Name))
-	}
-	c.indent--
-	c.writeln("};")
-	return nil
-}
-
-func (c *Compiler) compileFunExpr(fe *parser.FunExpr) (string, error) {
-	params := []string{}
-	for _, p := range fe.Params {
-		typ, err := c.compileType(p.Type)
-		if err != nil && typ == "auto" {
-			return "", err
-		}
-		params = append(params, fmt.Sprintf("%s %s", typ, p.Name))
-	}
-	ret := ""
-	if fe.Return != nil {
-		typ, err := c.compileType(fe.Return)
-		if err != nil && typ == "auto" {
-			return "", err
-		}
-		ret = " -> " + typ
-	}
-	if fe.ExprBody == nil {
-		return "", fmt.Errorf("lambda block body not supported")
-	}
-	body, err := c.compileExpr(fe.ExprBody)
-	if err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("([=](%s)%s { return %s; })", strings.Join(params, ", "), ret, body), nil
-}
-
-func (c *Compiler) compileStmt(s *parser.Statement) error {
+func (c *Compiler) compileStmt(st *parser.Statement) error {
 	switch {
-	case s.Let != nil:
+	case st.Let != nil:
 		c.writeIndent()
-		typ, err := c.compileType(s.Let.Type)
-		if err != nil && typ == "auto" {
-			return err
-		}
-		c.buf.WriteString(typ + " ")
-		c.buf.WriteString(s.Let.Name)
-		if s.Let.Value != nil {
-			c.buf.WriteString(" = ")
-			expr, err := c.compileExpr(s.Let.Value)
+		c.buf.WriteString("auto ")
+		c.buf.WriteString(st.Let.Name)
+		if st.Let.Value != nil {
+			val, err := c.compileExpr(st.Let.Value)
 			if err != nil {
 				return err
 			}
-			c.buf.WriteString(expr)
-		} else if s.Let.Type != nil {
-			if typ == "int" {
-				c.buf.WriteString(" = 0")
-			} else if typ == "string" {
-				c.buf.WriteString(" = \"\"")
-			}
+			c.buf.WriteString(" = ")
+			c.buf.WriteString(val)
 		}
 		c.buf.WriteString(";\n")
-	case s.Var != nil:
+	case st.Var != nil:
 		c.writeIndent()
-		typ, err := c.compileType(s.Var.Type)
-		if err != nil && typ == "auto" {
-			return err
-		}
-		c.buf.WriteString(typ + " ")
-		c.buf.WriteString(s.Var.Name)
-		if s.Var.Value != nil {
-			c.buf.WriteString(" = ")
-			expr, err := c.compileExpr(s.Var.Value)
+		c.buf.WriteString("auto ")
+		c.buf.WriteString(st.Var.Name)
+		if st.Var.Value != nil {
+			val, err := c.compileExpr(st.Var.Value)
 			if err != nil {
 				return err
 			}
-			c.buf.WriteString(expr)
-		} else if s.Var.Type != nil {
-			if typ == "int" {
-				c.buf.WriteString(" = 0")
-			} else if typ == "string" {
-				c.buf.WriteString(" = \"\"")
-			}
+			c.buf.WriteString(" = ")
+			c.buf.WriteString(val)
 		}
 		c.buf.WriteString(";\n")
-	case s.Assign != nil:
+	case st.Assign != nil:
 		c.writeIndent()
-		rhs, err := c.compileExpr(s.Assign.Value)
+		rhs, err := c.compileExpr(st.Assign.Value)
 		if err != nil {
 			return err
 		}
-		if len(s.Assign.Index) > 0 && len(s.Assign.Field) == 0 {
-			lhs := s.Assign.Name
-			for i := 0; i < len(s.Assign.Index)-1; i++ {
-				idx, err := c.compileExpr(s.Assign.Index[i].Start)
-				if err != nil {
-					return err
-				}
-				lhs = fmt.Sprintf("%s[%s]", lhs, idx)
-			}
-			idx, err := c.compileExpr(s.Assign.Index[len(s.Assign.Index)-1].Start)
-			if err != nil {
-				return err
-			}
-			c.buf.WriteString(fmt.Sprintf("%s[%s] = %s;\n", lhs, idx, rhs))
-		} else if len(s.Assign.Index) == 0 && len(s.Assign.Field) == 0 {
-			c.buf.WriteString(s.Assign.Name + " = " + rhs + ";\n")
-		} else {
-			return fmt.Errorf("complex assignment not supported")
-		}
-	case s.Return != nil:
+		c.buf.WriteString(st.Assign.Name)
+		c.buf.WriteString(" = ")
+		c.buf.WriteString(rhs)
+		c.buf.WriteString(";\n")
+	case st.Return != nil:
 		c.writeIndent()
-		expr, err := c.compileExpr(s.Return.Value)
+		val, err := c.compileExpr(st.Return.Value)
 		if err != nil {
 			return err
 		}
-		c.buf.WriteString("return " + expr + ";\n")
-	case s.Expr != nil:
+		c.buf.WriteString("return ")
+		c.buf.WriteString(val)
+		c.buf.WriteString(";\n")
+	case st.Expr != nil:
 		c.writeIndent()
-		expr, err := c.compileExpr(s.Expr.Expr)
+		val, err := c.compileExpr(st.Expr.Expr)
 		if err != nil {
 			return err
 		}
-		c.buf.WriteString(expr + ";\n")
-	case s.If != nil:
-		return c.compileIf(s.If)
-	case s.While != nil:
-		cond, err := c.compileExpr(s.While.Cond)
+		c.buf.WriteString(val)
+		c.buf.WriteString(";\n")
+	case st.If != nil:
+		return c.compileIf(st.If)
+	case st.While != nil:
+		cond, err := c.compileExpr(st.While.Cond)
 		if err != nil {
 			return err
 		}
 		c.writeIndent()
 		c.buf.WriteString("while(" + cond + ") {\n")
 		c.indent++
-		for _, b := range s.While.Body {
+		for _, b := range st.While.Body {
 			if err := c.compileStmt(b); err != nil {
 				return err
 			}
 		}
 		c.indent--
 		c.writeln("}")
-	case s.For != nil:
-		return c.compileFor(s.For)
-	case s.Break != nil:
+	case st.For != nil:
+		return c.compileFor(st.For)
+	case st.Break != nil:
 		c.writeIndent()
 		c.writeln("break;")
-	case s.Continue != nil:
+	case st.Continue != nil:
 		c.writeIndent()
 		c.writeln("continue;")
 	default:
-		return fmt.Errorf("unsupported statement at %v", s.Pos)
+		return fmt.Errorf("unsupported statement at %v", st.Pos)
 	}
 	return nil
 }
@@ -288,8 +189,8 @@ func (c *Compiler) compileIf(st *parser.IfStmt) error {
 	c.writeIndent()
 	c.buf.WriteString("if(" + cond + ") {\n")
 	c.indent++
-	for _, s := range st.Then {
-		if err := c.compileStmt(s); err != nil {
+	for _, b := range st.Then {
+		if err := c.compileStmt(b); err != nil {
 			return err
 		}
 	}
@@ -299,8 +200,8 @@ func (c *Compiler) compileIf(st *parser.IfStmt) error {
 	if st.Else != nil {
 		c.buf.WriteString(" else {\n")
 		c.indent++
-		for _, s := range st.Else {
-			if err := c.compileStmt(s); err != nil {
+		for _, b := range st.Else {
+			if err := c.compileStmt(b); err != nil {
 				return err
 			}
 		}
@@ -362,19 +263,15 @@ func (c *Compiler) compileExpr(e *parser.Expr) (string, error) {
 		if err != nil {
 			return "", err
 		}
-		out := left
+		result := left
 		for _, op := range e.Binary.Right {
 			rhs, err := c.compilePostfix(op.Right)
 			if err != nil {
 				return "", err
 			}
-			if op.Op == "in" {
-				out = fmt.Sprintf("([&](const auto& __c,const auto& __e){ using C=decay_t<decltype(__c)>; if constexpr(is_same_v<C,string>) return __c.find(__e) != string::npos; else if constexpr(has_key_type<C>::value) return __c.find(__e) != __c.end(); else return find(__c.begin(), __c.end(), __e) != __c.end(); })(%s,%s)", rhs, out)
-			} else {
-				out = fmt.Sprintf("(%s %s %s)", out, op.Op, rhs)
-			}
+			result = fmt.Sprintf("(%s %s %s)", result, op.Op, rhs)
 		}
-		return out, nil
+		return result, nil
 	}
 	return "", fmt.Errorf("unsupported expression at %v", e.Pos)
 }
@@ -389,6 +286,7 @@ func (c *Compiler) compileUnary(u *parser.Unary) (string, error) {
 	}
 	return val, nil
 }
+
 func (c *Compiler) compilePostfix(pf *parser.PostfixExpr) (string, error) {
 	target, err := c.compilePrimary(pf.Target)
 	if err != nil {
@@ -396,25 +294,8 @@ func (c *Compiler) compilePostfix(pf *parser.PostfixExpr) (string, error) {
 	}
 	expr := target
 	for _, op := range pf.Ops {
-		if op.Index != nil {
-			if op.Index.Colon != nil {
-				start, err := c.compileExpr(op.Index.Start)
-				if err != nil {
-					return "", err
-				}
-				end, err := c.compileExpr(op.Index.End)
-				if err != nil {
-					return "", err
-				}
-				expr = fmt.Sprintf("([&](const auto& __v){ using V=decay_t<decltype(__v)>; if constexpr(is_same_v<V,string>) return __v.substr(%s, %s-%s); else return vector<typename V::value_type>(__v.begin()+%s, __v.begin()+%s); })(%s)", start, end, start, start, end, expr)
-			} else {
-				idx, err := c.compileExpr(op.Index.Start)
-				if err != nil {
-					return "", err
-				}
-				expr = fmt.Sprintf("%s[%s]", expr, idx)
-			}
-		} else if op.Call != nil {
+		switch {
+		case op.Call != nil:
 			args := []string{}
 			for _, a := range op.Call.Args {
 				s, err := c.compileExpr(a)
@@ -423,67 +304,34 @@ func (c *Compiler) compilePostfix(pf *parser.PostfixExpr) (string, error) {
 				}
 				args = append(args, s)
 			}
-			if expr == "print" && len(args) == 1 {
-				expr = fmt.Sprintf("([&](const auto& __v){ if constexpr(has_size<decay_t<decltype(__v)>>::value){ for(size_t i=0;i<__v.size();++i){ if(i) cout<<' '; cout<<__v[i]; } cout<<endl; } else { cout<<__v<<endl; } })(%s)", args[0])
-			} else if expr == "print" {
-				e := "([&](){ cout"
-				for _, a := range args {
-					e += " << " + a
-				}
-				e += " << endl; })()"
-				expr = e
-			} else if expr == "len" && len(args) == 1 {
-				if strings.HasPrefix(args[0], "\"") {
-					expr = fmt.Sprintf("((int)string(%s).size())", args[0])
-				} else {
-					expr = fmt.Sprintf("((int)%s.size())", args[0])
-				}
-			} else if expr == "count" && len(args) == 1 {
-				expr = fmt.Sprintf("((int)%s.size())", args[0])
-			} else if expr == "min" && len(args) == 1 {
-				expr = fmt.Sprintf("(*min_element(%s.begin(), %s.end()))", args[0], args[0])
-			} else if expr == "max" && len(args) == 1 {
-				expr = fmt.Sprintf("(*max_element(%s.begin(), %s.end()))", args[0], args[0])
-			} else if expr == "str" && len(args) == 1 {
-				expr = fmt.Sprintf("to_string(%s)", args[0])
-			} else if expr == "append" && len(args) == 2 {
-				expr = fmt.Sprintf("([&](auto __v){ __v.push_back(%s); return __v; })(%s)", args[1], args[0])
-			} else if expr == "sum" && len(args) == 1 {
-				expr = fmt.Sprintf("([&](const auto& __v){ typename decay_t<decltype(__v)>::value_type __s{}; for(const auto& __x: __v) __s += __x; return __s; })(%s)", args[0])
-			} else if expr == "avg" && len(args) == 1 {
-				expr = fmt.Sprintf("([&](const auto& __v){ typename decay_t<decltype(__v)>::value_type __s{}; for(const auto& __x: __v) __s += __x; return __v.empty()?0:static_cast<double>(__s)/__v.size(); })(%s)", args[0])
-			} else if expr == "substring" && len(args) == 3 {
-				expr = fmt.Sprintf("%s.substr(%s, %s - %s)", args[0], args[1], args[2], args[1])
-			} else if expr == "values" && len(args) == 1 {
-				expr = fmt.Sprintf("([&](const auto& __m){ vector<decltype(__m.begin()->second)> __v; for(const auto& __kv: __m) __v.push_back(__kv.second); return __v; })(%s)", args[0])
-			} else if strings.HasSuffix(expr, ".contains") && len(args) == 1 {
-				base := strings.TrimSuffix(expr, ".contains")
-				expr = fmt.Sprintf("(%s.find(%s) != string::npos)", base, args[0])
+			if expr == "print" {
+				expr = fmt.Sprintf("([&](){ cout << %s << endl; })()", strings.Join(args, " << \" \" << "))
 			} else {
 				expr = fmt.Sprintf("%s(%s)", expr, strings.Join(args, ", "))
 			}
-		} else if op.Field != nil {
-			expr = fmt.Sprintf("%s.%s", expr, op.Field.Name)
-		} else if op.Cast != nil {
-			typ, err := c.compileType(op.Cast.Type)
+		case op.Index != nil:
+			idx, err := c.compileExpr(op.Index.Start)
 			if err != nil {
 				return "", err
 			}
-			if typ == "int" {
-				expr = fmt.Sprintf("stoi(%s)", expr)
-			} else {
-				expr = fmt.Sprintf("(%s)(%s)", typ, expr)
-			}
-		} else {
+			expr = fmt.Sprintf("%s[%s]", expr, idx)
+		case op.Field != nil:
+			expr = fmt.Sprintf("%s.%s", expr, op.Field.Name)
+		default:
 			return "", fmt.Errorf("unsupported postfix op")
 		}
 	}
 	return expr, nil
 }
+
 func (c *Compiler) compilePrimary(p *parser.Primary) (string, error) {
 	switch {
-	case p.Lit != nil:
-		return c.compileLiteral(p.Lit)
+	case p.Selector != nil:
+		name := p.Selector.Root
+		for _, t := range p.Selector.Tail {
+			name += "." + t
+		}
+		return name, nil
 	case p.Call != nil:
 		args := []string{}
 		for _, e := range p.Call.Args {
@@ -494,63 +342,12 @@ func (c *Compiler) compilePrimary(p *parser.Primary) (string, error) {
 			args = append(args, s)
 		}
 		fn := p.Call.Func
-		if fn == "print" && len(args) == 1 {
-			return fmt.Sprintf("([&](const auto& __v){ if constexpr(has_size<decay_t<decltype(__v)>>::value){ for(size_t i=0;i<__v.size();++i){ if(i) cout<<' '; cout<<__v[i]; } cout<<endl; } else { cout<<__v<<endl; } })(%s)", args[0]), nil
-		}
 		if fn == "print" {
-			out := "([&](){ cout"
-			for _, a := range args {
-				out += " << " + a
-			}
-			out += " << endl; })()"
-			return out, nil
-		}
-		if fn == "len" && len(args) == 1 {
-			if strings.HasPrefix(args[0], "\"") {
-				return fmt.Sprintf("((int)string(%s).size())", args[0]), nil
-			}
-			return fmt.Sprintf("((int)%s.size())", args[0]), nil
-		}
-		if fn == "count" && len(args) == 1 {
-			return fmt.Sprintf("((int)%s.size())", args[0]), nil
-		}
-		if fn == "min" && len(args) == 1 {
-			return fmt.Sprintf("(*min_element(%s.begin(), %s.end()))", args[0], args[0]), nil
-		}
-		if fn == "max" && len(args) == 1 {
-			return fmt.Sprintf("(*max_element(%s.begin(), %s.end()))", args[0], args[0]), nil
-		}
-		if fn == "str" && len(args) == 1 {
-			return fmt.Sprintf("to_string(%s)", args[0]), nil
-		}
-		if fn == "append" && len(args) == 2 {
-			return fmt.Sprintf("([&](auto __v){ __v.push_back(%s); return __v; })(%s)", args[1], args[0]), nil
-		}
-		if fn == "sum" && len(args) == 1 {
-			return fmt.Sprintf("([&](const auto& __v){ typename decay_t<decltype(__v)>::value_type __s{}; for(const auto& __x: __v) __s += __x; return __s; })(%s)", args[0]), nil
-		}
-		if fn == "avg" && len(args) == 1 {
-			return fmt.Sprintf("([&](const auto& __v){ typename decay_t<decltype(__v)>::value_type __s{}; for(const auto& __x: __v) __s += __x; return __v.empty()?0:static_cast<double>(__s)/__v.size(); })(%s)", args[0]), nil
-		}
-		if fn == "substring" && len(args) == 3 {
-			return fmt.Sprintf("%s.substr(%s, %s - %s)", args[0], args[1], args[2], args[1]), nil
-		}
-		if fn == "values" && len(args) == 1 {
-			return fmt.Sprintf("([&](const auto& __m){ vector<decltype(__m.begin()->second)> __v; for(const auto& __kv: __m) __v.push_back(__kv.second); return __v; })(%s)", args[0]), nil
+			return fmt.Sprintf("([&](){ cout << %s << endl; })()", strings.Join(args, " << \" \" << ")), nil
 		}
 		return fn + "(" + strings.Join(args, ", ") + ")", nil
-	case p.FunExpr != nil:
-		return c.compileFunExpr(p.FunExpr)
-	case p.Struct != nil:
-		fields := []string{}
-		for _, f := range p.Struct.Fields {
-			s, err := c.compileExpr(f.Value)
-			if err != nil {
-				return "", err
-			}
-			fields = append(fields, s)
-		}
-		return p.Struct.Name + "{" + strings.Join(fields, ", ") + "}", nil
+	case p.Lit != nil:
+		return c.compileLiteral(p.Lit)
 	case p.List != nil:
 		elems := []string{}
 		for _, e := range p.List.Elems {
@@ -561,42 +358,6 @@ func (c *Compiler) compilePrimary(p *parser.Primary) (string, error) {
 			elems = append(elems, s)
 		}
 		return "vector<int>{" + strings.Join(elems, ", ") + "}", nil
-	case p.Map != nil:
-		pairs := []string{}
-		keyType := "string"
-		valType := "int"
-		for i, it := range p.Map.Items {
-			k, err := c.compileExpr(it.Key)
-			if err != nil {
-				return "", err
-			}
-			v, err := c.compileExpr(it.Value)
-			if err != nil {
-				return "", err
-			}
-			if i == 0 {
-				if strings.HasPrefix(k, "std::string") || strings.HasPrefix(k, "\"") {
-					keyType = "string"
-				} else {
-					keyType = "int"
-				}
-				if strings.HasPrefix(v, "std::string") || strings.HasPrefix(v, "\"") {
-					valType = "string"
-				} else if strings.HasPrefix(v, "unordered_map") {
-					valType = v[:strings.Index(v, ">")+1]
-				} else {
-					valType = "int"
-				}
-			}
-			pairs = append(pairs, fmt.Sprintf("{%s, %s}", k, v))
-		}
-		return fmt.Sprintf("unordered_map<%s, %s>{%s}", keyType, valType, strings.Join(pairs, ", ")), nil
-	case p.Selector != nil:
-		name := p.Selector.Root
-		for _, t := range p.Selector.Tail {
-			name += "." + t
-		}
-		return name, nil
 	case p.Group != nil:
 		return c.compileExpr(p.Group)
 	default:
@@ -618,27 +379,7 @@ func (c *Compiler) compileLiteral(l *parser.Literal) (string, error) {
 		return "false", nil
 	}
 	if l.Str != nil {
-		return fmt.Sprintf("std::string(%q)", *l.Str), nil
-	}
-	if l.Null {
-		return "nullptr", nil
+		return fmt.Sprintf("\"%s\"", *l.Str), nil
 	}
 	return "", fmt.Errorf("unknown literal")
-}
-
-func (c *Compiler) compileType(t *parser.TypeRef) (string, error) {
-	if t == nil {
-		return "auto", nil
-	}
-	if t.Simple != nil {
-		switch *t.Simple {
-		case "int":
-			return "int", nil
-		case "string":
-			return "string", nil
-		default:
-			return *t.Simple, nil
-		}
-	}
-	return "auto", fmt.Errorf("unsupported type")
 }
